+++ dla_cnn/desi/io.py
@@ -2,6 +2,13 @@
 
 from dla_cnn.data_model import Sightline
 
+
+def read_mock_spectrum(id, dla_catalog):
+    sightline = Sightline(id=id)
+
+    """ Rest of this needs building """
+
+
 def read_sightline(id):
     """
     Read Sightline from hard drive
