+++ dla_cnn/sdss/Model_v4.py
